+++ test/ReadonlyArray.ts
@@ -1,5 +1,4 @@
 import {
-  length,
   elemFlipped,
   none,
   join,
@@ -58,19 +57,6 @@ import { NonEmptyArray } from "fp-ts/NonEmptyArray"
 import * as T from "fp-ts/These"
 
 describe("Array", () => {
-  describe("length", () => {
-    const f = length
-
-    it("returns length of array", () => {
-      expect(f([])).toBe(0)
-      expect(f([1, 2, 3])).toBe(3)
-    })
-
-    it("returned value is always non-negative", () => {
-      fc.assert(fc.property(fc.array(fc.anything()), xs => f(xs) >= 0))
-    })
-  })
-
   describe("elemFlipped", () => {
     const f = elemFlipped(N.Eq)
 
@@ -113,14 +99,14 @@ describe("Array", () => {
           const countDelims = flow(
             split(""),
             A.filter(c => c === delim),
-            length,
+            A.size,
           )
 
           const countDelimsA = flow(A.map(countDelims), concatAll(N.MonoidSum))
 
           return (
             countDelims(f(xs)) ===
-            countDelimsA(xs) + max(N.Ord)(0, length(xs) - 1)
+            countDelimsA(xs) + max(N.Ord)(0, A.size(xs) - 1)
           )
         }),
       )
@@ -232,13 +218,13 @@ describe("Array", () => {
         fc.property(
           fc.array(fc.anything(), 1, 5) as fc.Arbitrary<NonEmptyArray<unknown>>,
           fc.integer(0, 10),
-          (xs, i) => O.isSome(f(i)(xs)(xs)) === i <= length(xs),
+          (xs, i) => O.isSome(f(i)(xs)(xs)) === i <= A.size(xs),
         ),
       )
 
       fc.assert(
         fc.property(fc.array(fc.string(), 1, 20), xs =>
-          expect(pipe(g(xs), O.map(length))).toEqual(O.some(length(xs) + 2)),
+          expect(pipe(g(xs), O.map(A.size))).toEqual(O.some(A.size(xs) + 2)),
         ),
       )
     })
@@ -265,7 +251,7 @@ describe("Array", () => {
 
     it("never returns a larger array", () => {
       fc.assert(
-        fc.property(fc.array(fc.integer()), xs => f(xs).length <= length(xs)),
+        fc.property(fc.array(fc.integer()), xs => f(xs).length <= A.size(xs)),
       )
     })
 
@@ -357,7 +343,7 @@ describe("Array", () => {
         fc.property(
           fc.array(fc.anything()),
           fc.array(fc.anything()),
-          (xs, ys) => f(xs)(ys).length === length(xs) * ys.length,
+          (xs, ys) => f(xs)(ys).length === A.size(xs) * ys.length,
         ),
       )
     })
@@ -428,7 +414,7 @@ describe("Array", () => {
       fc.assert(
         fc.property(
           fc.array(fc.anything()),
-          xs => !f(length(xs) + 1)(xs).length,
+          xs => !f(A.size(xs) + 1)(xs).length,
         ),
       )
     })
@@ -452,7 +438,7 @@ describe("Array", () => {
         fc.property(
           fc.array(fc.anything()),
           fc.integer(1, 100),
-          (xs, n) => f(n)(xs).length === max(N.Ord)(0, length(xs) - n + 1),
+          (xs, n) => f(n)(xs).length === max(N.Ord)(0, A.size(xs) - n + 1),
         ),
       )
     })
@@ -486,7 +472,7 @@ describe("Array", () => {
       fc.assert(
         fc.property(
           fc.array(fc.integer()),
-          xs => A.filter(p)(xs).length + f(xs).length === length(xs),
+          xs => A.filter(p)(xs).length + f(xs).length === A.size(xs),
         ),
       )
     })
@@ -502,7 +488,7 @@ describe("Array", () => {
         fc.property(
           fc.array(fc.anything(), 1, n),
           fc.integer(0, n - 1),
-          (xs, i) => i >= length(xs) || expect(f(i)(i)(xs)).toEqual(O.some(xs)),
+          (xs, i) => i >= A.size(xs) || expect(f(i)(i)(xs)).toEqual(O.some(xs)),
         ),
       )
     })
@@ -510,7 +496,7 @@ describe("Array", () => {
     it("returns None if source is out of bounds", () => {
       fc.assert(
         fc.property(fc.array(fc.anything()), xs =>
-          expect(f(length(xs))(0)(xs)).toEqual(O.none),
+          expect(f(A.size(xs))(0)(xs)).toEqual(O.none),
         ),
       )
     })
@@ -518,7 +504,7 @@ describe("Array", () => {
     it("returns None if target is out of bounds", () => {
       fc.assert(
         fc.property(fc.array(fc.anything()), xs =>
-          expect(f(0)(length(xs))(xs)).toEqual(O.none),
+          expect(f(0)(A.size(xs))(xs)).toEqual(O.none),
         ),
       )
     })
@@ -550,7 +536,7 @@ describe("Array", () => {
           (xs, i, j) =>
             pipe(
               f(i)(j)(xs),
-              O.exists(ys => length(ys) === n),
+              O.exists(ys => A.size(ys) === n),
             ),
         ),
       )
@@ -579,7 +565,7 @@ describe("Array", () => {
         fc.property(
           fc.array(fc.anything(), 1, n),
           fc.integer(0, n - 1),
-          (xs, i) => i >= length(xs) || expect(f(i)(i)(xs)).toEqual(O.some(xs)),
+          (xs, i) => i >= A.size(xs) || expect(f(i)(i)(xs)).toEqual(O.some(xs)),
         ),
       )
     })
@@ -587,7 +573,7 @@ describe("Array", () => {
     it("returns None if source is out of bounds", () => {
       fc.assert(
         fc.property(fc.array(fc.anything()), xs =>
-          expect(f(0)(length(xs))(xs)).toEqual(O.none),
+          expect(f(0)(A.size(xs))(xs)).toEqual(O.none),
         ),
       )
     })
@@ -595,7 +581,7 @@ describe("Array", () => {
     it("returns None if target is out of bounds", () => {
       fc.assert(
         fc.property(fc.array(fc.anything()), xs =>
-          expect(f(length(xs))(0)(xs)).toEqual(O.none),
+          expect(f(A.size(xs))(0)(xs)).toEqual(O.none),
         ),
       )
     })
@@ -627,7 +613,7 @@ describe("Array", () => {
           (xs, i, j) =>
             pipe(
               f(i)(j)(xs),
-              O.exists(ys => length(ys) === n),
+              O.exists(ys => A.size(ys) === n),
             ),
         ),
       )
@@ -811,7 +797,7 @@ describe("Array", () => {
       fc.assert(
         fc.property(
           fc.array(fc.array(fc.anything())),
-          xs => length(A.flatten(xs)) === length(A.flatten(f(xs))),
+          xs => A.size(A.flatten(xs)) === A.size(A.flatten(f(xs))),
         ),
       )
     })